GHOST-PROTOCOL/services/brain/requirements.txt
fastapi
uvicorn
google-generativeai
redis
python-dotenv
psycopg2-binary
sqlalchemy
pydantic
httpx
